requirements.txt
# AI/ML/DL Core Libraries - Python 3.13 + Apple Silicon Compatible
# tensorflow>=2.13.0  # NOT COMPATIBLE with Python 3.13 yet
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Deep Learning & Neural Networks - Python 3.13 Compatible
torch>=2.0.0  # PyTorch supports Python 3.13
torchvision>=0.15.0
torchaudio>=2.0.0
# transformers>=4.30.0  # May not support Python 3.13 yet
# pytorch-lightning>=2.0.0  # May not support Python 3.13 yet

# Financial Data Analysis
yfinance>=0.2.18
pandas-datareader>=0.10.0
# quantlib>=1.30  # Commented out - complex C++ dependencies
zipline-reloaded>=3.0.0
backtrader>=1.9.78
# ta-lib>=0.4.25  # Commented out - requires system-level TA-Lib installation
pyfolio>=0.9.2
empyrical>=0.5.5
alphalens-reloaded>=0.4.0

# Time Series Analysis
statsmodels==0.14.1
arch==6.2.0
pmdarima==2.0.4
prophet==1.1.5
tslearn==0.6.2
tsfresh==0.20.2

# Market Data & APIs
alpha-vantage==2.3.1
iexfinance==0.5.0
quandl==3.7.0
yahoo-fin==0.8.9.1
ccxt==4.1.77
websocket-client==1.7.0
requests==2.31.0
aiohttp==3.9.1

# Real-time Data Processing
kafka-python==2.0.2
redis==5.0.1
celery==5.3.4
dramatiq==1.15.0
apache-airflow==2.8.1

# Database Connectors
influxdb-client==1.39.0
psycopg2-binary==2.9.9
pymongo==4.6.1
sqlalchemy==2.0.25
alembic==1.13.1
clickhouse-driver==0.2.6

# NLP & Sentiment Analysis
nltk==3.8.1
spacy==3.7.2
textblob==0.17.1
vaderSentiment==3.3.2
transformers[sentencepiece]==4.36.2
newspaper3k==0.2.8
beautifulsoup4==4.12.2

# Mathematical & Statistical Computing
sympy==1.12
cvxpy==1.4.1
cvxopt==1.3.2
pymc==5.10.0
pystan==3.8.0
arviz==0.17.0

# Risk Management & Portfolio Optimization
riskfolio-lib==5.1.0
pypfopt==1.5.5
bt==0.2.9
vectorbt==0.25.2
pyportfolioopt==1.5.5

# Technical Analysis
# TA-Lib>=0.4.25  # Commented out - requires system TA-Lib installation
pandas-ta>=0.3.14
stockstats>=0.6.0
finta>=1.3

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
bokeh==3.3.4
dash==2.16.1
streamlit==1.29.0

# Web Framework & APIs
fastapi==0.108.0
uvicorn==0.25.0
django==4.2.8
django-rest-framework==3.14.0
flask==3.0.0
gunicorn==21.2.0

# Async & Concurrency
aiofiles>=23.0.0
# asyncio is built-in to Python 3.7+
# concurrent-futures is built-in to Python 3.2+

# Data Validation & Serialization
pydantic==2.5.2
marshmallow==3.20.2
cerberus==1.3.5

# Configuration & Environment
python-decouple==3.8
pydantic-settings==2.1.0
dynaconf==3.2.4

# Logging & Monitoring
loguru==0.7.2
structlog==23.2.0
prometheus-client==0.19.0
sentry-sdk==1.39.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.2
pytest-cov==4.1.0
pytest-mock==3.12.0
hypothesis==6.96.1

# Code Quality
black==23.12.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
bandit==1.7.5

# Jupyter & Development
jupyter==1.0.0
jupyterlab==4.0.9
ipython==8.18.1
notebook==7.0.6

# Cloud & Infrastructure
boto3==1.34.25
google-cloud-storage==2.10.0
azure-storage-blob==12.19.0
docker==7.0.0
kubernetes==28.1.0

# Encryption & Security
cryptography==41.0.8
pyotp==2.9.0
bcrypt==4.1.2
pyjwt==2.8.0

# Alternative Data Sources
tweepy==4.14.0
praw==7.7.1
feedparser==6.0.10
scrapy==2.11.0

# Performance & Profiling
line-profiler>=4.0.0
memory-profiler>=0.60.0
py-spy>=0.3.12
# cProfile is built-in to Python

# Utilities
python-dateutil==2.8.2
pytz==2023.3.post1
tqdm==4.66.1
joblib==1.3.2
click==8.1.7
rich==13.7.0
typer==0.9.0

# Financial Calculations
mibian==0.1.3
py-vollib==1.0.1
quantlib-python==1.32
business-calendar==0.3.0

# Regex & Text Processing
regex==2023.12.25
python-docx==1.1.0
openpyxl==3.1.2
xlrd==2.0.1

# HTTP & Networking
httpx==0.26.0
urllib3==2.1.0
certifi==2023.11.17

# Data Structures & Algorithms
networkx>=3.0.0
python-igraph>=0.10.0
# graph-tool>=2.50  # Commented out - complex dependencies on macOS

# GPU Computing (Optional - for ML acceleration - Linux/Windows only)
# cupy-cuda12x>=13.0.0  # Commented out - CUDA not available on macOS
# rapids-cudf>=23.0.0   # Commented out - RAPIDS not available on macOS

# Environment & System
psutil>=5.9.0 yfinance
alpaca-trade-api
vollib
transformers
torch
newsapi-python
redis
